portfolio_lab/requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.11.0
matplotlib>=3.8.0
PyYAML>=6.0.0
jupyter>=1.0.0
notebook>=7.0.0
